# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para conversion de PDF, IIIF Presentation v2 (con v3 compatible), IIIF Image API, administracion y migracion. Las rutas recomendadas usan /api/v1 y los endpoints legacy se mantienen por compatibilidad temporal.",
        "title": "IIIF PDF Server API",
        "contact": {},
        "version": "2.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/documents/{id}/ocr/jobs": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Inicia una generación OCR híbrida, exhaustiva o solo OCR. Requiere sesión administrativa.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Crear trabajo OCR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Configuración del trabajo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateOCRJobRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/services.OCRJob"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/ocr/jobs/{job_id}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Consultar progreso OCR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del trabajo",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.OCRJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/ocr/jobs/{job_id}/cancel": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Cancelar trabajo OCR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del trabajo",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.OCRJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/documents/{id}/ocr": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Obtener estado OCR de un documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.OCRDocumentSummary"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/{id}/ocr/pages/{page}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Obtener OCR y bounding boxes de una página",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.OCRPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/{id}/ocr/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Buscar texto OCR dentro de un documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Texto a buscar",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Máximo de resultados",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Desplazamiento",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/ocr/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Buscar texto OCR por proyecto, tenant o documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Texto a buscar",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Proyecto",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "document_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Máximo de resultados",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Desplazamiento",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "detalle interno"
                },
                "error": {
                    "type": "string",
                    "example": "login requerido"
                }
            }
        },
...
        "services.CreateOCRJobRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "type": "boolean"
                },
                "language_mode": {
                    "type": "string"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mode": {
                    "type": "string"
                }
            }
        },
        "services.OCRDocumentSummary": {
            "type": "object",
            "properties": {
                "active_generation": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "failed_pages": {
                    "type": "integer"
                },
                "indexed_pages": {
                    "type": "integer"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "project_key": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tenant_key": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "services.OCRJob": {
            "type": "object",
            "properties": {
                "cancel_requested": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "current_page": {
                    "type": "integer"
                },
                "document_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "failed_pages": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "generation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language_mode": {
                    "type": "string"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mode": {
                    "type": "string"
                },
                "processed_pages": {
                    "type": "integer"
                },
                "project_key": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tenant_key": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "services.OCRPage": {
            "type": "object",
            "properties": {
                "canvas_v2": {
                    "type": "string"
                },
                "canvas_v3": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "engine": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "generation": {
                    "type": "string"
                },
                "geometry_status": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "native_text": {
                    "type": "string"
                },
                "ocr_text": {
                    "type": "string"
                },
                "page_number": {
                    "type": "integer"
                },
                "schema_version": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "words": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.OCRWord"
                    }
                }
            }
        },
        "services.OCRWord": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "height": {
                    "type": "integer"
                },
                "left": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "top": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "project_iiif_session",
            "in": "cookie"
        }
    }
}